contributors/adobe/submitted/shapes/shape-outside/values/shape-outside-shape-inherit-000.html

<!DOCTYPE html>
<html>
    <head>
        <title>Shape Outside Box Initial Value</title>
        <link rel="author" title="Adobe" href="http://html.adobe.com/">
        <link rel="author" title="Bear Travis" href="mailto:betravis@adobe.com">
        <link rel="help" href="http://dev.w3.org/csswg/css-shapes/#propdef-shape-outside">
        <meta name="assert" content="Shape-outside takes its default value of auto when assigned the 'initial' value">
        <meta name="flags" content="DOM">
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="../resources/parsing-utils.js"></script>
    </head>
    <body>
        <div id="log"></div>
        <script type="text/javascript">
            test(function() {
                var outer = document.createElement("div");
                var inner = document.createElement("div");
                outer.appendChild(inner);
                outer.style.setProperty("shape-outside", "content-box");
                inner.style.setProperty("shape-outside", "inherit");
                document.body.appendChild(outer);
                var inline = inner.style.getPropertyValue("shape-outside");
                var style = getComputedStyle(inner);
                var computed = style.getPropertyValue("shape-outside");
                document.body.removeChild(outer);
                assert_equals(inline, "inherit");
                assert_equals(computed, "content-box");
            }, "shape-outside can be assigned 'inherit' value");
        </script>
    </body>
</html>
